♻️ (server): Pagination code refactoring
