+++ apps/server/src/pagination/paginationUserBookmarks/domain/models/paginationUserBookmarks.filter.ts
@@ -1,4 +1,5 @@
 export interface PaginationUserBookmarksFilter {
   normalizedTag?: string;
   interestId?: string;
+  myUserBookmark?: boolean;
 }
